add api account

+++ app_api/app/http/api/account/admin/services.py
@@ -7,41 +7,19 @@ from fastapi import (
 from sqlalchemy import (
 	func, 
 	desc,
+	or_,
 )
 from app import models as mdl
 from . import schemas as sch
 from app.config import cfg
 
 
-def get_item_by_id(db: DB, model, id):
-	item = db.session.execute(
-		db.select(model).filter_by(id=id)
-	).scalar()
-	if not item:
-		raise HTTPException(
-			status_code=status.HTTP_404_NOT_FOUND, 
-			detail='Account not found'
-		)
-	return item
-
-def check_account_uniqueness_by(db: DB, email: str, username: str):
-	account = db.session.execute(
-		db.select(mdl.Account).filter(or_(
-			mdl.Account.username == username, 
-			mdl.Account.email == email
-		))
-	).scalar()
-	if account is not None:
-		raise HTTPException(
-			status_code=status.HTTP_400_BAD_REQUEST,
-			detail='Username or email not unique',
-		)
 
 
 def get_accounts(db: DB, skip: int = 0, limit: int = cfg.items_in_list):
 	select_accounts = db.select(
 			mdl.Account.id, mdl.Account.email, mdl.Account.username,
-			mdl.Profile.first_name, mdl.Profile.last_name, mdl.Profile.sex,
+			mdl.Profile.first_name, mdl.Profile.last_name, mdl.Profile.female,
 			func.count(mdl.Article.id).label('articles_count')
 		).\
 		outerjoin(mdl.Account.articles).outerjoin(mdl.Account.profile).\
@@ -56,7 +34,7 @@ def get_accounts(db: DB, skip: int = 0, limit: int = cfg.items_in_list):
 def get_account(db: DB, account_id: int):
 	select_account = db.select(
 			mdl.Account.id, mdl.Account.email, mdl.Account.username,
-			mdl.Profile.first_name, mdl.Profile.last_name, mdl.Profile.sex,
+			mdl.Profile.first_name, mdl.Profile.last_name, mdl.Profile.female,
 			func.count(mdl.Article.id).label('articles_count')
 		).\
 		filter_by(id=account_id).\
@@ -72,11 +50,20 @@ def get_account(db: DB, account_id: int):
 
 
 def create_account(db: DB, account_data: sch.AccountInCreate):
-	check_account_uniqueness_by(db, account_data.email, account_data.username)
+	account = mdl.Account.get_first_item_by_filter(
+		db, _or=True, email=account_data.email, username=account_data.username
+	)
+	if account is not None:
+		raise HTTPException(
+			status_code=status.HTTP_400_BAD_REQUEST,
+			detail='Username or email not unique',
+		) from None
 
 	new_account = mdl.Account(
 		email=account_data.email,
 		username=account_data.username,
+		is_blocked=account_data.is_blocked,
+		is_shown=account_data.is_shown,
 		password=mdl.Account.get_hashed_password(account_data.password),
 	)
 	db.session.add(new_account)
@@ -86,31 +73,37 @@ def create_account(db: DB, account_data: sch.AccountInCreate):
 	new_profile = mdl.Profile(
 		first_name=account_data.profile.first_name,
 		last_name=account_data.profile.first_name,
-		sex=True if account_data.profile.sex == 'female' else False,
-		account_id=new_account.id
+		female=account_data.profile.female,
+		account_id=new_account.id,
 	)
 	db.session.add(new_profile)
 	db.session.commit()
 	return new_account
 
 
+
 def update_account(db: DB, account_id: int, account_data: sch.AccountInCreate):
-	account = get_item_by_id(db, mdl.Account, account_id)
+	account = mdl.Account.get_first_item_by_filter(db, id=account_id)
+	if account is None:
+		raise HTTPException(
+			status_code=status.HTTP_404_NOT_FOUND, 
+			detail='Account not found'
+		) from None
 	
-	account.email=account_data.email,
-	account.username=account_data.username,
-	account.password=mdl.Account.get_hashed_password(account_data.password),
+	account.email=account_data.email
+	account.username=account_data.username
+	account.is_blocked=account_data.is_blocked
+	account.is_shown=account_data.is_shown
+	account.password=mdl.Account.get_hashed_password(account_data.password)
 	db.session.add(account)
 	db.session.commit()
 	db.session.refresh(account)
 
-	profile = db.session.execute(
-		db.select(mdl.Profile).filter_by(account_id=account.id)
-	).scalar()
-	if profile:
-		profile.first_name=account_data.profile.first_name,
-		profile.last_name=account_data.profile.first_name,
-		profile.sex=True if account_data.profile.sex == 'female' else False,
+	profile = mdl.Profile.get_first_item_by_filter(db, account_id=account.id)
+	if profile is not None:
+		profile.first_name=account_data.profile.first_name
+		profile.last_name=account_data.profile.first_name
+		female=account_data.profile.female
 		db.session.add(profile)
 		db.session.commit()
 	
@@ -118,11 +111,17 @@ def update_account(db: DB, account_id: int, account_data: sch.AccountInCreate):
 
 
 def delete_account(db: DB, account_id: int):
-	account = get_item_by_id(db, mdl.Account, account_id)
-	db.session.delete(account)
-	profile = db.session.execute(
-		db.select(mdl.Profile).filter_by(account_id=account.id)
-	).scalar()
-	if profile:
+	account = mdl.Account.get_first_item_by_filter(db, id=account_id)
+	if not account:
+		raise HTTPException(
+			status_code=status.HTTP_404_NOT_FOUND, 
+			detail='Account not found'
+		) from None
+	profile = mdl.Profile.get_first_item_by_filter(db, account_id=account.id)
+	if profile is not None:
 		db.session.delete(profile)
+	# account.delete(synchronize_session=False)
+	db.session.delete(account)
 	db.session.commit()
+
+
